Update Calculadora_EfrenGutierrez.java
Cambiamos las variables numero1 y numero 2 de int a float

+++ Calculadora_EfrenGutierrez/src/main/java/hoja6/calculadora_efrengutierrez/Calculadora_EfrenGutierrez.java
@@ -14,7 +14,7 @@ public class Calculadora_EfrenGutierrez {
 
     static Scanner scanner = new Scanner(System.in);
     static int opcion = -1; //opción del menú
-    static int numero1 = 0, numero2 = 0; //Variables de entrada
+    static float numero1 = 0, numero2 = 0; //Variables de entrada
 
     public static void main(String[] args) {
         System.out.println("Bienvenido a tu calculadora");
@@ -59,10 +59,10 @@ public class Calculadora_EfrenGutierrez {
     //Método para recoger las variables de entrada
     public static void pideNumeros() {
         System.out.println("Introduce el primer número:");
-        numero1 = scanner.nextInt();
+        numero1 = scanner.nextFloat();
 
         System.out.println("Introduce el segundo número:");
-        numero2 = scanner.nextInt();
+        numero2 = scanner.nextFloat();
 
     }
 }
